Crate intl_memoizer

This crate contains a memoizer for internationalization formatters. Often it is expensive (in terms of performance and memory) to construct a formatter, but then relatively cheap to run the format operation.

The IntlMemoizer is the main struct that creates a per-locale IntlLangMemoizer.

Modules

concurrent

Contains thread-safe variants.

Structs

IntlLangMemoizer

The IntlLangMemoizer can memoize multiple constructed internationalization formatters, and their configuration for a single locale. For instance, given “en-US”, a memorizer could retain 3 DateTimeFormat instances, and a PluralRules.

IntlMemoizer

IntlMemoizer is designed to handle lazily-initialized references to internationalization formatters.

Traits

Memoizable

The trait that needs to be implemented for each intl formatter that needs to be memoized.